+++ mycodo/config.py
@@ -13,7 +13,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '7.1.7'
-ALEMBIC_VERSION = 'ed21c36670f4'
+ALEMBIC_VERSION = 'a8341ac0d779'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.