+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
  - Fix calculating Output Usage
 
+### Features
+
+ - Add new energy usage/cost analysis based on amperage measurements (See [Energy Usage](https://github.com/kizniche/Mycodo/blob/master/mycodo-manual.md#energy-usage) in the manual)
+
 
 ## 7.1.7 (2019-02-02)